+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * `EMP-10852` API changes to `MediaPlayback.currentTime`. Property renamed to `MediaPlayback.playheadPosition` which better reflects the actual position of the playback.
 * `SessionShift` protocol has been reamed to `StartTime` as that better reflects its purpose on `Tech` level.
 * Improvements in how `HLSNative` handles `StartTime`.
+* Event callbacks now return `Player` instead of `Tech` directly.
+* Seekable and Bbuffered timeranges now return `[CMTimeRange]`
 
 ## 0.73.0